Map markers are selectable with map dialogue.
In this commit, we make map markers selectable
where a pop up appears when the marker is
clicked. Currently, this contains the stop
name which sadly also needs to be pulled from
the Metlink API with another call. This means
that there can be a delay from seeing the pop
up to reading the name of the stop.

This is also implemented with yet another state
prop. There are quite a lot on the screen sadly.
This will likely need to be tidied up a bit in
future revisions.