Creates Generic GoogleMapWidget.tsx to handle all mapping.
In this commit, we extract duplicate code from both components
that contained Google Map widgets. This functionality has been
brought to GoogleMapWidget.tsx. To assist with this, TypeScript
has been leveraged to define some classes that can be used to
help parse and verify the data with confidence. The
MetlinkStopMap.tsx and MetlinkServiceView.tsx have been refactored
to utilise this new component.

Note: In this commit, we broke the stop names for the service view.