GoogleMapWidget.tsx can now handle Markers with no names.
In this commit, we add support for StopMarker's to have no name.
To handle this, when a GoogleMapWidget.tsx is downloaded, so too
is the data from Transitfeed containing all of the stop names. In
the situation that a StopMarker's name is null then the widget will
attempt to find a name from the downloaded data. In the situation
where this does not yield any results. Then 'stopCode - unknown' will
be returned.