Slightly changes site mappings to isolate stop/ and service/
In this commit, we slightly tweak how our site mappings look
to isolate the stop and service perspectives into their own
separate page urls. This is part of a bigger vision
regarding site structure and how we implement our back
buttons. We also add a SITE_STRUCTURE.md that outlines the
sites mappings as well as a basic break down of each pages
functionality.