Fixes bug where info windows unreadable in dark theme.
In this commit we fix a known bug where info windows
are unreadable when the user has selected dark theme.
To fix this, we added a custom CSS argument to
override the existing GoogleMaps CSS. This commit also
adds and references any missing CSS files to save time
down the track.