Fixes "e.g." functionality.
All strings are treated like new paragraphs. The rule before was that any string that CONTAINED "e.g." would be treated as a new line rather than a new paragraph. Rather, now, any line that START WITH "e.g." will be treated as a new line, any string that contains "e.g." will be treated like a normal string. This allows for a more diverse setup of text.