Current non-working attempt.
In this commit, we add a segment to the Settings page where
the user can switch between using auto, light or dark theme.

This segment changes the local storage to reflect the users
preference. Unfortunately, I cannot seem to translate to a
visual change in the browser. You can see my attempts within
App.tsx.