Adds generic ListComponent for all stop and service list items.
In this commit, we add a generic component to act as a wrapper
for IonItems in lists. This aims to make the UI more consistent
as well as clean up the code in the respective classes. We also,
tweak the search algorithm to account for both the badge and the
name of the stop/service. The changes make place to the SearchTab,
the MetlinkStopTable and the MetlinkSavedList.