Adds functionality to toggle between hours and time.
In this commit, we add the functionality to toggle between hours and
time. This is done by adding an optional prop to the ListComponent
that contains time. This is rendered in non-bold on the line below
the list entry. The user can select whether to show hours or time
based off an IonToggle. (The entire IonItem is clickable, not just
the toggle).