Moves GoogleMap controls to bottom of screen.
In this commit, we play with the GoogleMap
API typescript definitions and alter the
options to move the UI elements to the bottom
of the screen. Unfortunately, although WebStorm
could find the definitions, and there were no
compile errors in the console the page would
not load due to 'google' not being found. Gah,
these map to enums in the definitions so I have
just hardcoded these for now. Just makes
this slightly less readable.