SearchBar now on top of map, on select prints coordinates.
In this commit, we reshuffle the MapTab structure to bring
the search bar onto the map. This is the only fix I could
find to enable our search results to appear on top of the
map without stretching the IonTitle. We also add support
for clicking the elements such where the destination
coordinates are printed to the console.