Extracts Firebase authentication logic and adds a user waiting screen.
In this commit we do two notable things. Firstly, we extract any Firebase
related code from the AccountLoginPerspective and AccountSignupPerspective
perspectives into the Firebase. This is an aim help contain the Firebase
logic into the one library. We also add a user waiting screen to indicate
to the user the context is still loading and remove any sudden redirects
such as from the login perspective to the account perspective. To implement
this, we needed to change how we implemented the UserContext to add a pre
state (such that we can identify whether or not the context has mounted).