Implements caching of stop and service data.
In this commit, we implement local caching of stop / service data.
This information will be used if the StorageManager is unable to
obtain a fresh set of data. The error message will be propagated to
the user via an error card, which has now been abstracted from the
account perspectives.